AS-IS) Board NoOffset 적용 TO-BE) NoOffset 확인

+++ src/main/java/com/spring/sharepod/v1/controller/BoardRestController.java
@@ -34,7 +34,7 @@ public class BoardRestController {
     //8번 API 릴스 동영상 get 하는 방식 (구현 완료)
     // stress_test
     @GetMapping("/board/video")
-    public ResponseEntity<AllVideo> getVideo(@RequestParam(value = "startNum", defaultValue = "0") int startNum) {
+    public ResponseEntity<AllVideo> getVideo(@RequestParam(value = "startNum", defaultValue = "0") Long startNum) {
         //limit 안 들어오면 5로 고정
         //Long validLimitCount = boardValidator.DefaultLimitCount(limitCount);
